refactor(fast_io): split io_uring/registered_buffers/tests.rs by concern
The post-#4310 tests.rs at 882 lines exceeded the 650-line cap. Split
into a tests/ directory grouped by concern, each file well under cap:

- tests/mod.rs (39) - shared try_ring / try_group helpers and module decls.
- tests/registry.rs (113) - page size, validation guards, slot allocator.
- tests/submit.rs (179) - submit_read/write_fixed_batch helpers.
- tests/stats.rs (141) - acquire/miss counters and snapshot semantics.
- tests/status.rs (49) - try_new_with_status outcome reporting.
- tests/drop_contract.rs (309) - Drop semantics and constrained-env coverage
  for the PR #4022 / task #2118 fixed-buffer invariants audit.

The try_ring / try_group helpers replace the repeated match-or-return
boilerplate that handled environments where io_uring is unavailable
(CI sandboxes, seccomp, kernel < 5.6).

No behavioural change; test coverage is preserved.
